backend/requirements.txt
phe
supabase
bcrypt
flask
python-dotenv